Splitted ViewingGrid into GameView and ViewingGrid.
To make it make more sense to create a copy of ViewingGrid on the client
side, I made it work against the interfaces I_Tile and I_IslandGrid.
These interfaces contains methods that can be implemented on the
client's side of things. The class ViewingGrid can therefore be used in
both the server and the client.